+++ client/app/scripts/controllers/products/products-review-controller.js
@@ -1,5 +1,5 @@
 (function() {
-    'use strict';
+	'use strict';
 
 	/**
 	 * @ngdoc function
@@ -9,32 +9,71 @@
 	 * Controller of the gemStoreApp
 	 */
 	angular.module('gemStoreApp')
-		.controller("ReviewCtrl", ['$scope', '$resource', 'productsService', function ($scope, $resource, productsService) {
+		.controller('ReviewCtrl', ['$scope', '$resource', 'Restangular', 'productsService', function ($scope, $resource, Restangular, productsService) {
 			this.review = {};
 
 			// Method for adding a review
 			this.addReview = function(product){
 
-				var url = '/'+ product._id + '/reviews';
+				//console.log(product._id);
 
-				console.log("What is url? " + url);
-
-				console.log('posting a review to the front end!!');
+				//console.log('posting a review to the front end!!');
 				this.review.createdOn = Date.now();
 
-				console.log("What is product: " + angular.toJson(product));
+				//************
+				//RESTANGULR METHOD
 
-				product.reviews.push(this.review);
+				// RestangularProvider.setRestangularFields({
+				// 	id: "_id"
+				// });
 
-				console.log("What is product: " + angular.toJson(product));
+				var productReview = Restangular.all('/products/' + product._id + '/reviews');
 
-				productsService.update('products/reviews', product, function() {
-                    console.log("Save successfull!!!");
-                    $modalInstance.close();
-                });
+				var updatedProduct = product;
+
+				//console.log('updatedProduct: ' + angular.toJson(updatedProduct));
+				//console.log('product: ' + angular.toJson(product));
+
+				console.log('New review: ' + angular.toJson(this.review));
+
+				//Pushing the review to the front-end
+				//product.reviews.push(this.review);
+
+				//console.log('What is product after adding a review to Front-end: ' + angular.toJson(product));
+
+				//productsService.save('/products', product, function() {
+				//productReview.post(product).then(function(newResource){
+				//updatedProduct.post(productReview, product).then(function(newResource){)
+				productReview.post(product).then(function(newResource){
+					//productsService.update(productReview, product);
+					console.log('product: ' + angular.toJson(product));
+					console.log('What is newResource?' + angular.toJson(newResource));
+
+					//product.reviews.push(newResource);
+					//console.log('What is product after adding a review: ' + angular.toJson(product));
+				});
+
+				//************
+				//NG Resource Method
+
+				//var url = 'localhost:3000/products/'+ product._id + '/reviews';
+				//var url = 'localhost:3000/products/55afbcf000318fed33038ae0/reviews';
+				//var url = 'product';
+
+				//localhost:3000/products/55afbcf000318fed33038ae0/reviews
+
+				//console.log('What is url? ' + url);
+
+				// console.log('New review: ' + angular.toJson(this.review));
+
+				// $http.post('/products' + id + '/comments', comment)
+				// productsService.save({id: product._id +'/reviews'},product, function() {
+				// 	console.log('am I am getting here?')
+				// 	console.log('Save successfull!!!');
+				// });
 
 				// productsService.save(product, function() {
-				// 	console.log("Update the server?");
+				// 	console.log('Update the server?');
 				// 	//product.$update();
 				// 	productsService.$update(product);
 				// });